Got rid of InkNode, as any Ink classes are not used. Added require.js as a separate file.
JS files are included with requireJS, so the CoffeeScript files have to be built first.

Filled the stories with text.

Added audio to the page (Tsjaikovsky's Valse Des Fleurs, isn't that nice?).

Using the WaitForImages jQuery plugin to load all images before init'ing the page.

Also using HashChange jQuery plugin for handling hash events, as hash used for the stories and scenes.

Created a README, for later reference.